astronomer-course/requirements.txt
psycopg2-binary
apache-airflow-providers-postgres
pandas
sqlalchemy